requirements.txt
# pip install -r requirements.txt

# ─── Núcleo do Algoritmo ───────────────────────────────────────────────
numpy==2.2.6                # Operações matriciais (matriz de distâncias, fitness)
scipy==1.15.3               # Otimização matemática auxiliar

# ─── Algoritmos Evolutivos ─────────────────────────────────────────────
deap==1.4.3                 # Framework para Algoritmos Genéticos (GA, GP, ES)

# ─── Dados Geoespaciais ────────────────────────────────────────────────
haversine==2.9.0            # Distância entre coordenadas GPS (fórmula Haversine)
shapely==2.1.2              # Geometria: polígonos de zonas de exclusão aérea
geojson==3.2.0              # Leitura de zonas proibidas em formato GeoJSON

# ─── Comunicação com Arduino ───────────────────────────────────────────
pymavlink==2.4.49           # Protocolo MAVLink para envio de waypoints ao drone
pyserial==3.5               # Comunicação serial USB/Bluetooth com Arduino

# ─── Meteorologia em Tempo Real ────────────────────────────────────────
requests==2.32.5            # Consulta a APIs externas (clima, elevação)
python-dotenv==1.2.2        # Variáveis de ambiente (chaves de API, portas seriais)

# ─── Visualização de Rotas ─────────────────────────────────────────────
matplotlib==3.10.8           # Plots 2D das rotas calculadas
folium==0.20.0               # Mapas interativos HTML com rotas sobrepostas
networkx==3.4.2              # Representação e análise do grafo de rotas

# ─── Simulação ─────────────────────────────────────────────────────────
simpy==4.1.1                 # Simulação de eventos discretos (voos simultâneos)

# ─── Testes ────────────────────────────────────────────────────────────
pytest==9.0.2               # Framework de testes
pytest-cov==7.0.0            # Cobertura de testes
pytest-asyncio==0.23.7
httpx==0.27.0


# ── Banco de Dados PostgreSQL + PostGIS ────────────────────────────────────────
sqlalchemy==2.0.31
asyncpg==0.31.0
alembic==1.13.2

# ── Servidor FastAPI ───────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
#validar depois
pydantic==2.7.4